+++ hutool-json/src/test/java/cn/hutool/json/JSONObjectTest.java
@@ -15,6 +15,7 @@ import cn.hutool.core.io.resource.ResourceUtil;
 import cn.hutool.core.lang.Console;
 import cn.hutool.core.util.StrUtil;
 import cn.hutool.json.test.bean.JSONBean;
+import cn.hutool.json.test.bean.ResultDto;
 import cn.hutool.json.test.bean.Seq;
 import cn.hutool.json.test.bean.TokenAuthResponse;
 import cn.hutool.json.test.bean.TokenAuthWarp2;
@@ -118,7 +119,11 @@ public class JSONObjectTest {
 
 	@Test
 	public void toBeanNullStrTest() {
-		JSONObject json = JSONUtil.createObj().put("strValue", "null").put("intValue", 123).put("beanValue", "null").put("list", JSONUtil.createArray().put("a").put("b"));
+		JSONObject json = JSONUtil.createObj()//
+				.put("strValue", "null")//
+				.put("intValue", 123)//
+				.put("beanValue", "null")//
+				.put("list", JSONUtil.createArray().put("a").put("b"));
 
 		TestBean bean = json.toBean(TestBean.class);
 		// 当JSON中为字符串"null"时应被当作字符串处理
@@ -195,6 +200,17 @@ public class JSONObjectTest {
 		Assert.assertNotNull(result);
 		Assert.assertEquals("tokenTest", result.getToken());
 		Assert.assertEquals("测试用户1", result.getUserId());
+	}
+	
+	/**
+	 * 泛型对象中的泛型参数如果未定义具体类型，按照JSON处理<br>
+	 * 此处用于测试获取泛型类型实际类型错误导致的空指针问题
+	 */
+	@Test
+	public void toBeanTest7() {
+		String jsonStr = " {\"result\":{\"phone\":\"15926297342\",\"appKey\":\"e1ie12e1ewsdqw1\",\"secret\":\"dsadadqwdqs121d1e2\",\"message\":\"hello world\"},\"code\":100,\"message\":\"validate message\"}";
+		ResultDto<?> dto = JSONUtil.toBean(jsonStr, ResultDto.class);
+		Assert.assertEquals("validate message", dto.getMessage());
 	}
 	
 	@Test